Fixed bugs in 2.0b1.
Implemented a dedicated preferences controller that's responsible for all user prefs, default prefs, the Preferences window, and preference upgrades.  Removed some old preferences that are never used and confusing.

Implemented an application delegate that's responsible for routing messages around.  Moved responsibility for the About window to the application delegate.

Split the Profile Manager into its own nib and completely redid the interface to it.  Changed the default order of encodings so that the fastest is now the default.  Future-proofed the encodings so that we can change the order or add/remove new encodings easily.

The connection dialog form is now disabled if there is no connection specified.

Cleaned up the work-flow for using command line options.  CLI invocations can now optionally specify a profile name.


git-svn-id: https://chicken.svn.sourceforge.net/svnroot/chicken/trunk@184 1fd050d8-3d9c-4ecf-acd6-3d06a6b47c07